iota_json_rpc_api

Trait ExtendedApiServer

pub trait ExtendedApiServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn get_epochs<'life0, 'async_trait>(
        &'life0 self,
        cursor: Option<BigInt<u64>>,
        limit: Option<usize>,
        descending_order: Option<bool>,
    ) -> Pin<Box<dyn Future<Output = RpcResult<EpochPage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_epoch_metrics<'life0, 'async_trait>(
        &'life0 self,
        cursor: Option<BigInt<u64>>,
        limit: Option<usize>,
        descending_order: Option<bool>,
    ) -> Pin<Box<dyn Future<Output = RpcResult<EpochMetricsPage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_current_epoch<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<EpochInfo>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_network_metrics<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<NetworkMetrics>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_move_call_metrics<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<MoveCallMetrics>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_latest_address_metrics<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_checkpoint_address_metrics<'life0, 'async_trait>(
        &'life0 self,
        checkpoint: u64,
    ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_all_epoch_address_metrics<'life0, 'async_trait>(
        &'life0 self,
        descending_order: Option<bool>,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<AddressMetrics>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_total_transactions<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<BigInt<u64>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_participation_metrics<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<ParticipationMetrics>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the ExtendedApi RPC API.

Required Methods

fn get_epochs<'life0, 'async_trait>( &'life0 self, cursor: Option<BigInt<u64>>, limit: Option<usize>, descending_order: Option<bool>, ) -> Pin<Box<dyn Future<Output = RpcResult<EpochPage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a list of epoch info. Exclusively served by the indexer.

fn get_epoch_metrics<'life0, 'async_trait>( &'life0 self, cursor: Option<BigInt<u64>>, limit: Option<usize>, descending_order: Option<bool>, ) -> Pin<Box<dyn Future<Output = RpcResult<EpochMetricsPage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a list of epoch metrics, which is a subset of epoch info. Exclusively served by the indexer.

fn get_current_epoch<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<EpochInfo>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return current epoch info. Exclusively served by the indexer.

fn get_network_metrics<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<NetworkMetrics>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return Network metrics. Exclusively served by the indexer.

fn get_move_call_metrics<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<MoveCallMetrics>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return move call metrics. Exclusively served by the indexer.

fn get_latest_address_metrics<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Address related metrics. Exclusively served by the indexer.

fn get_checkpoint_address_metrics<'life0, 'async_trait>( &'life0 self, checkpoint: u64, ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Address related metrics. Exclusively served by the indexer.

fn get_all_epoch_address_metrics<'life0, 'async_trait>( &'life0 self, descending_order: Option<bool>, ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<AddressMetrics>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Address related metrics. Exclusively served by the indexer.

fn get_total_transactions<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<BigInt<u64>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the total number of transactions. Exclusively served by the indexer.

fn get_participation_metrics<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<ParticipationMetrics>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the participation metrics. Participation is defined as the total number of unique addresses that have delegated stake in the current epoch. Includes both staked and timelocked staked IOTA. Exclusively served by the indexer.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.